// src/grpc/typescript/api.ts
// @generated by protobuf-ts 2.9.4 with parameter long_type_string,optimize_code_size
// @generated from protobuf file "api.proto" (package "api", syntax proto3)
// tslint:disable
import { MessageType } from "@protobuf-ts/runtime";
import { Any } from "./google/protobuf/any";
/**
 * @generated from protobuf message api.IBackendGRPC
 */
export interface IBackendGRPC {
    /**
     * @generated from protobuf field: repeated string error = 1;
     */
    error: string[];
    /**
     * @generated from protobuf field: string message = 2;
     */
    message: string;
    /**
     * @generated from protobuf field: bool status = 3;
     */
    status: boolean;
    /**
     * @generated from protobuf field: string data = 4;
     */
    data: string;
    /**
     * @generated from protobuf field: int32 code = 5;
     */
    code: number;
}
/**
 * @generated from protobuf message api.Meta
 */
export interface Meta {
    /**
     * @generated from protobuf field: int32 current = 1;
     */
    current: number;
    /**
     * @generated from protobuf field: int32 pageSize = 2;
     */
    pageSize: number;
    /**
     * @generated from protobuf field: int32 totalPage = 3;
     */
    totalPage: number;
    /**
     * @generated from protobuf field: int32 totalItem = 4;
     */
    totalItem: number;
}
/**
 * @generated from protobuf message api.IModelPaginate
 */
export interface IModelPaginate {
    /**
     * @generated from protobuf field: api.Meta meta = 1;
     */
    meta?: Meta;
    /**
     * @generated from protobuf field: repeated google.protobuf.Any result = 2;
     */
    result: Any[];
}
// @generated message type with reflection information, may provide speed optimized methods
class IBackendGRPC$Type extends MessageType<IBackendGRPC> {
    constructor() {
        super("api.IBackendGRPC", [
            { no: 1, name: "error", kind: "scalar", repeat: 2 /*RepeatType.UNPACKED*/, T: 9 /*ScalarType.STRING*/ },
            { no: 2, name: "message", kind: "scalar", T: 9 /*ScalarType.STRING*/ },
            { no: 3, name: "status", kind: "scalar", T: 8 /*ScalarType.BOOL*/ },
            { no: 4, name: "data", kind: "scalar", T: 9 /*ScalarType.STRING*/ },
            { no: 5, name: "code", kind: "scalar", T: 5 /*ScalarType.INT32*/ }
        ]);
    }
}
/**
 * @generated MessageType for protobuf message api.IBackendGRPC
 */
export const IBackendGRPC = new IBackendGRPC$Type();
// @generated message type with reflection information, may provide speed optimized methods
class Meta$Type extends MessageType<Meta> {
    constructor() {
        super("api.Meta", [
            { no: 1, name: "current", kind: "scalar", T: 5 /*ScalarType.INT32*/ },
            { no: 2, name: "pageSize", kind: "scalar", T: 5 /*ScalarType.INT32*/ },
            { no: 3, name: "totalPage", kind: "scalar", T: 5 /*ScalarType.INT32*/ },
            { no: 4, name: "totalItem", kind: "scalar", T: 5 /*ScalarType.INT32*/ }
        ]);
    }
}
/**
 * @generated MessageType for protobuf message api.Meta
 */
export const Meta = new Meta$Type();
// @generated message type with reflection information, may provide speed optimized methods
class IModelPaginate$Type extends MessageType<IModelPaginate> {
    constructor() {
        super("api.IModelPaginate", [
            { no: 1, name: "meta", kind: "message", T: () => Meta },
            { no: 2, name: "result", kind: "message", repeat: 1 /*RepeatType.PACKED*/, T: () => Any }
        ]);
    }
}
/**
 * @generated MessageType for protobuf message api.IModelPaginate
 */
export const IModelPaginate = new IModelPaginate$Type();