// src/grpc/typescript/category-restaurant.client.ts
// @generated by protobuf-ts 2.9.4 with parameter long_type_string,optimize_code_size
// @generated from protobuf file "category-restaurant.proto" (package "CategoryRestaurantProto", syntax proto3)
// tslint:disable
import type { RpcTransport } from "@protobuf-ts/runtime-rpc";
import type { ServiceInfo } from "@protobuf-ts/runtime-rpc";
import { CategoryRestaurantServiceGprc } from "./category-restaurant";
import type { ReqFindCatResByArrId } from "./category-restaurant";
import { stackIntercept } from "@protobuf-ts/runtime-rpc";
import type { IBackendGRPC } from "./api";
import type { ReqFindOneCatRes } from "./category-restaurant";
import type { UnaryCall } from "@protobuf-ts/runtime-rpc";
import type { RpcOptions } from "@protobuf-ts/runtime-rpc";
/**
 * @generated from protobuf service CategoryRestaurantProto.CategoryRestaurantServiceGprc
 */
export interface ICategoryRestaurantServiceGprcClient {
    /**
     * @generated from protobuf rpc: findOneCatRes(CategoryRestaurantProto.ReqFindOneCatRes) returns (api.IBackendGRPC);
     */
    findOneCatRes(input: ReqFindOneCatRes, options?: RpcOptions): UnaryCall<ReqFindOneCatRes, IBackendGRPC>;
    /**
     * @generated from protobuf rpc: findCatResByArrId(CategoryRestaurantProto.ReqFindCatResByArrId) returns (api.IBackendGRPC);
     */
    findCatResByArrId(input: ReqFindCatResByArrId, options?: RpcOptions): UnaryCall<ReqFindCatResByArrId, IBackendGRPC>;
}
/**
 * @generated from protobuf service CategoryRestaurantProto.CategoryRestaurantServiceGprc
 */
export class CategoryRestaurantServiceGprcClient implements ICategoryRestaurantServiceGprcClient, ServiceInfo {
    typeName = CategoryRestaurantServiceGprc.typeName;
    methods = CategoryRestaurantServiceGprc.methods;
    options = CategoryRestaurantServiceGprc.options;
    constructor(private readonly _transport: RpcTransport) {
    }
    /**
     * @generated from protobuf rpc: findOneCatRes(CategoryRestaurantProto.ReqFindOneCatRes) returns (api.IBackendGRPC);
     */
    findOneCatRes(input: ReqFindOneCatRes, options?: RpcOptions): UnaryCall<ReqFindOneCatRes, IBackendGRPC> {
        const method = this.methods[0], opt = this._transport.mergeOptions(options);
        return stackIntercept<ReqFindOneCatRes, IBackendGRPC>("unary", this._transport, method, opt, input);
    }
    /**
     * @generated from protobuf rpc: findCatResByArrId(CategoryRestaurantProto.ReqFindCatResByArrId) returns (api.IBackendGRPC);
     */
    findCatResByArrId(input: ReqFindCatResByArrId, options?: RpcOptions): UnaryCall<ReqFindCatResByArrId, IBackendGRPC> {
        const method = this.methods[1], opt = this._transport.mergeOptions(options);
        return stackIntercept<ReqFindCatResByArrId, IBackendGRPC>("unary", this._transport, method, opt, input);
    }
}